--- src/model/Category.ts
@@ -1,14 +1,0 @@
-import { v4 as uuidV4 } from 'uuid'
-
-export class Category {
-    id?: string;
-    name: string;
-    description: string;
-    created_at: Date;
-
-    constructor() {
-        if (!this.id) {
-            this.id = uuidV4()
-        }
-    }
-}
