+++ src/modules/cars/model/Category.ts
@@ -1,0 +1,14 @@
+import { v4 as uuidV4 } from 'uuid'
+
+export class Category {
+  id?: string;
+  name: string;
+  description: string;
+  created_at: Date;
+
+  constructor() {
+    if (!this.id) {
+      this.id = uuidV4()
+    }
+  }
+}
