+++ src/modules/cars/model/Specification.ts
@@ -1,0 +1,16 @@
+import { v4 as uuidV4 } from 'uuid'
+
+class Specification {
+  id?: string;
+  name: string;
+  description: string;
+  created_at: Date;
+
+  constructor() {
+    if (!this.id) {
+      this.id = uuidV4()
+    }
+  }
+}
+
+export { Specification }
