+++ src/modules/cars/repositories/ICreateCategoryRepository.ts
@@ -1,14 +1,14 @@
 import { Category } from "../model/Category"
 
 interface ICreateCategoryDTO {
-    name: string,
-    description: string
+  name: string,
+  description: string
 }
 
 interface ICreateCategoryRepository {
-    findByName(name: string): Category,
-    create({ }: ICreateCategoryDTO): void,
-    list(): Category[]
+  findByName(name: string): Category,
+  create({ }: ICreateCategoryDTO): void,
+  list(): Category[]
 }
 
 export { ICreateCategoryDTO, ICreateCategoryRepository }
