+++ src/modules/cars/services/CreateSpecificationService.ts
@@ -1,0 +1,9 @@
+
+
+class CreateSpecificationService {
+  execute() {
+
+  }
+}
+
+export { CreateSpecificationService }
